requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
Flask-Login==0.6.3
Flask-SocketIO==5.1.1
cryptography==3.4.7
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.1.0
Werkzeug==2.3.7 